requirements.txt
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pyyaml>=6.0
openai>=1.0.0
